Update off net461 kinda everywhere
This commit is one possible solution to #96. Here, I'm simply dropping support
for net461 entirely, in favor of targeting `netstandard`. While I was doing
this, I bumped the `netstandard` version to 2.1, but this could easily be an
addition instead of a replacement. Similarly, I noted that the `global.json`
specified SDK version 3.0, but with a roll-forward directive for newest minor
version. Accordingly, I bumped to the newest SDK 3.1 release. These changes
meant I could remove the `mono` block from the Fleece solution, so I did.

These changes reach all through Fleece, and prompted the replacement of `net461`
with `netstandard2.1` (mostly in documentation) or `netcoreapp3.1` (mostly in
tests) in many places. While I was doing this, I cleaned whitespace and
indentation, and added a missing XML specifier to an fsproj file.

On my machine, the build and all tests are passing.